requirements-core.txt
# FranklinOps Core Dependencies
# Essential packages for running FranklinOps system

# ==================== FastAPI & Web Framework ====================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ==================== Async & HTTP ====================
httpx==0.25.2

# ==================== Cryptography & Security ====================
cryptography==41.0.7
pynacl==1.5.0

# ==================== Database ====================
sqlalchemy==2.0.23

# ==================== Data Processing ====================
pandas==2.1.3
numpy>=1.21.0
requests>=2.28.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
lxml>=4.9.0
pypdf
python-docx
openpyxl

# ==================== Logging ====================
loguru==0.7.2

# ==================== Environment & Config ====================
python-dotenv==1.0.0
pyyaml==6.0.1

# ==================== Utilities ====================
python-dateutil==2.8.2